+++ paxman/capabilities/UUID/rules/__init__.py
@@ -1,0 +1,1 @@
+"""UUID validation rules."""